Add absence logs to Calendar
